Refactoring Implementation and Design smells where needed
