+++ pacman-core/src/test/java/de/amr/games/pacman/test/PacManGameTest.java
@@ -46,7 +46,7 @@ public class PacManGameTest {
         var level = game.level().get();
         assertEquals(1, level.number());
         assertEquals(0, level.totalNumGhostsKilled());
-        assertEquals(0, level.numGhostsKilledByEnergizer());
+//        assertEquals(0, level.numGhostsKilledByEnergizer());
         assertEquals(0, level.cruiseElroyState());
     }
 
@@ -63,22 +63,22 @@ public class PacManGameTest {
     public void testGhostsCreatedAndInitialized() {
         game.level().ifPresent(level -> {
             var redGhost = level.ghost(GameModel.RED_GHOST);
-            assertEquals(-1, redGhost.killedIndex());
+//            assertEquals(-1, redGhost.killedIndex());
             assertNotEquals(Vector2f.ZERO, level.ghostRevivalPosition(redGhost.id()));
             assertNotEquals(Vector2i.ZERO, level.ghostScatterTarget(redGhost.id()));
 
             var pinkGhost = level.ghost(GameModel.PINK_GHOST);
-            assertEquals(-1, pinkGhost.killedIndex());
+//            assertEquals(-1, pinkGhost.killedIndex());
             assertNotEquals(Vector2f.ZERO, level.ghostRevivalPosition(pinkGhost.id()));
             assertNotEquals(Vector2i.ZERO, level.ghostScatterTarget(pinkGhost.id()));
 
             var cyanGhost = level.ghost(GameModel.CYAN_GHOST);
-            assertEquals(-1, cyanGhost.killedIndex());
+//            assertEquals(-1, cyanGhost.killedIndex());
             assertNotEquals(Vector2f.ZERO, level.ghostRevivalPosition(cyanGhost.id()));
             assertNotEquals(Vector2i.ZERO, level.ghostScatterTarget(cyanGhost.id()));
 
             var orangeGhost = level.ghost(GameModel.ORANGE_GHOST);
-            assertEquals(-1, orangeGhost.killedIndex());
+//            assertEquals(-1, orangeGhost.killedIndex());
             assertNotEquals(Vector2f.ZERO, level.ghostRevivalPosition(orangeGhost.id()));
             assertNotEquals(Vector2i.ZERO, level.ghostScatterTarget(orangeGhost.id()));
         });
@@ -139,10 +139,10 @@ public class PacManGameTest {
         assertEquals(0, GameController.it().credit());
     }
 
-    @Test(expected = IllegalArgumentException.class)
-    public void testIllegalKilledIndex() {
-        game.level().ifPresent(level -> level.ghost(GameModel.RED_GHOST).setKilledIndex(42));
-    }
+//    @Test(expected = IllegalArgumentException.class)
+//    public void testIllegalKilledIndex() {
+//        game.level().ifPresent(level -> level.ghost(GameModel.RED_GHOST).setKilledIndex(42));
+//    }
 
     @Test
     public void testLegalCruiseElroyState() {
