+++ pacman-ui-fx-2d/src/main/java/de/amr/games/pacman/ui/fx/ActionHandler.java
@@ -46,4 +46,5 @@ public interface ActionHandler {
     void cheatEnterNextLevel();
 
     void cheatKillAllEatableGhosts();
+    void toggle2D3D();
 }
