+++ pacman-ui-fx-2d/src/main/java/de/amr/games/pacman/ui/fx/PacManGames2dUI.java
@@ -843,4 +843,7 @@ public class PacManGames2dUI implements GameEventListener, GameSceneContext, Act
             voiceClipExecution.stop();
         }
     }
+
+    public void toggle2D3D(){
+    }
 }
