+++ pacman-ui-fx-3d/src/main/java/de/amr/games/pacman/ui/fx/v3d/entity/LivesCounter3D.java
@@ -20,6 +20,8 @@ import javafx.scene.shape.Cylinder;
 import javafx.scene.shape.DrawMode;
 import javafx.scene.transform.Rotate;
 import javafx.util.Duration;
+import de.amr.games.pacman.ui.fx.GameSceneContext;
+import static de.amr.games.pacman.ui.fx.v3d.PacManGames3dUI.*;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -70,6 +72,31 @@ public class LivesCounter3D extends Group {
 
         getChildren().addAll(standsGroup, light);
     }
+
+
+    public static void createLivesCounter3D(GameSceneContext context) {
+        var theme = context.theme();
+        LivesCounter3D livesCounter3D = new LivesCounter3D(
+                theme.get("livescounter.entries"),
+                theme.color("livescounter.pillar.color"),
+                theme.get("livescounter.pillar.height"),
+                theme.color("livescounter.plate.color"),
+                theme.get("livescounter.plate.thickness"),
+                theme.get("livescounter.plate.radius"),
+                theme.color("livescounter.light.color"));
+        livesCounter3D.setTranslateX(2 * TS);
+        livesCounter3D.setTranslateY(2 * TS);
+        livesCounter3D.drawModePy.bind(PY_3D_DRAW_MODE);
+        for (int i = 0; i < livesCounter3D.maxLives(); ++i) {
+            var pac3D = switch (context.gameVariant()) {
+                case MS_PACMAN -> Pac3D.createMsPacMan3D(context.theme(), null, theme.get("livescounter.pac.size"));
+                case    PACMAN -> Pac3D.createPacMan3D(context.theme(), null,  theme.get("livescounter.pac.size"));
+            };
+            livesCounter3D.addItem(pac3D, true);
+        }
+}
+
+
 
     public void addItem(Pac3D pac3D, boolean lookRight) {
         int x = pac3DList.size() * 2 * TS;
